Moved GetCurrentState() to correct class
