+++ conwaylib/BoardExtensions.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Drawing;
 using static System.Math;
 
 namespace ConwayLib
@@ -122,6 +124,43 @@ namespace ConwayLib
             }
 
             return newBoard;
+        }
+
+        public static GameState GetCurrentState(this Board board, DensityOption option)
+        {
+            var state = new GameState
+            {
+                Width = board.Width,
+                Height = board.Height
+            };
+            switch (option)
+            {
+                case DensityOption.Dense:
+                    state.DenseData = new bool[state.Height][];
+                    for (int y = 0; y < board.Height; y++)
+                    {
+                        state.DenseData[y] = new bool[state.Width];
+                        for (int x = 0; x < board.Width; x++)
+                        {
+                            state.DenseData[y][x] = board.Cell(x, y);
+                        }
+                    }
+                    break;
+                case DensityOption.Sparse:
+                    List<Point> tempPoints = new();
+                    for (int y = 0; y < board.Height; y++)
+                    {
+                        for (int x = 0; x < board.Width; x++)
+                        {
+                            if (board.Cell(x, y))
+                            {
+                                tempPoints.Add(new Point(x, y));
+                            }
+                        }
+                    }
+                    break;
+            }
+            return state;
         }
     }
 }
